rustc: Move the privacy pass to its own crate
